Switched int counts to longs instead.

+++ Schema/src/binary/reader/BinaryReaderExtensions_Known.cs
@@ -35,7 +35,7 @@ public static partial class BinaryReaderExtensions {
     => br.ReadFloatArrayStructs_(dst);
 
   [MethodImpl(MethodImplOptions.AggressiveInlining)]
-  public static Vector2[] ReadVector2s(this IBinaryReader br, int count) {
+  public static Vector2[] ReadVector2s(this IBinaryReader br, long count) {
     var array = new Vector2[count];
     br.ReadVector2s(array);
     return array;
@@ -51,7 +51,7 @@ public static partial class BinaryReaderExtensions {
     => br.ReadFloatArrayStructs_(dst);
 
   [MethodImpl(MethodImplOptions.AggressiveInlining)]
-  public static Vector3[] ReadVector3s(this IBinaryReader br, int count) {
+  public static Vector3[] ReadVector3s(this IBinaryReader br, long count) {
     var array = new Vector3[count];
     br.ReadVector3s(array);
     return array;
@@ -67,7 +67,7 @@ public static partial class BinaryReaderExtensions {
     => br.ReadFloatArrayStructs_(dst);
 
   [MethodImpl(MethodImplOptions.AggressiveInlining)]
-  public static Vector4[] ReadVector4s(this IBinaryReader br, int count) {
+  public static Vector4[] ReadVector4s(this IBinaryReader br, long count) {
     var array = new Vector4[count];
     br.ReadVector4s(array);
     return array;
@@ -83,7 +83,7 @@ public static partial class BinaryReaderExtensions {
     => br.ReadFloatArrayStructs_(dst);
 
   [MethodImpl(MethodImplOptions.AggressiveInlining)]
-  public static Matrix3x2[] ReadMatrix3x2s(this IBinaryReader br, int count) {
+  public static Matrix3x2[] ReadMatrix3x2s(this IBinaryReader br, long count) {
     var array = new Matrix3x2[count];
     br.ReadMatrix3x2s(array);
     return array;
@@ -99,7 +99,7 @@ public static partial class BinaryReaderExtensions {
     => br.ReadFloatArrayStructs_(dst);
 
   [MethodImpl(MethodImplOptions.AggressiveInlining)]
-  public static Matrix4x4[] ReadMatrix4x4s(this IBinaryReader br, int count) {
+  public static Matrix4x4[] ReadMatrix4x4s(this IBinaryReader br, long count) {
     var array = new Matrix4x4[count];
     br.ReadMatrix4x4s(array);
     return array;
@@ -116,7 +116,8 @@ public static partial class BinaryReaderExtensions {
     => br.ReadFloatArrayStructs_(dst);
 
   [MethodImpl(MethodImplOptions.AggressiveInlining)]
-  public static Quaternion[] ReadQuaternions(this IBinaryReader br, int count) {
+  public static Quaternion[]
+      ReadQuaternions(this IBinaryReader br, long count) {
     var array = new Quaternion[count];
     br.ReadQuaternions(array);
     return array;
